Edited and project directory.
-Deleted various jar files present inside the directory.
- renamed Spring-dependency to flyspring.
- removed unnecessary files in starter project.
