--- Spring-dependency/autoroute/src/main/java/com/flyspring/autoroute/annotations/PathVariableAnnotation.java
@@ -1,14 +1,0 @@
-package com.flyspring.autoroute.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface PathVariableAnnotation {
-
-    public String[] name() ;
-    
-}
