--- Spring-dependency/autoroute/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-com.flyspring.autoroute.RouterFunctionConfig
-com.flyspring.autoroute.AuthConfig
