+++ flyspring/autoroute/src/main/java/com/flyspring/autoroute/PropertiesLoader.java
@@ -1,0 +1,19 @@
+package com.flyspring.autoroute;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
+
+public class PropertiesLoader {
+
+    public static Properties loadProperties(String resourceFileName) throws IOException {
+        Properties configuration = new Properties();
+        InputStream inputStream = PropertiesLoader.class
+          .getClassLoader()
+          .getResourceAsStream(resourceFileName);
+        configuration.load(inputStream);
+        inputStream.close();
+        return configuration;
+    }
+    
+}
